+++ frontend/app/protected/dashboard/components/recent-task.tsx
@@ -1,82 +1,35 @@
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
+import { Task } from "@/services/dto/task-dto";
 
-export function RecentTask() {
+interface RecenTaskProps {
+  tasks: Task[];
+}
+
+export function RecentTask({ tasks }: RecenTaskProps) {
   return (
     <div className="space-y-8">
-      <div className="flex items-center">
-        <Avatar className="h-9 w-9">
-          <AvatarImage src="/avatars/01.png" alt="Avatar" />
-          <AvatarFallback>OM</AvatarFallback>
-        </Avatar>
-        <div className="ml-4 space-y-1">
-          <p className="text-sm font-medium leading-none">Olivia Martin</p>
-          <p className="text-sm text-muted-foreground">
-            olivia.martin@email.com
-          </p>
-        </div>
-        <div className="ml-auto font-medium">
-          <p className="text-sm font-medium leading-none">Onboarding MyBank</p>
-          <p className="text-sm text-muted-foreground text-end">OMB-200</p>
-        </div>
-      </div>
-      <div className="flex items-center">
-        <Avatar className="flex h-9 w-9 items-center justify-center space-y-0 border">
-          <AvatarImage src="/avatars/02.png" alt="Avatar" />
-          <AvatarFallback>JL</AvatarFallback>
-        </Avatar>
-        <div className="ml-4 space-y-1">
-          <p className="text-sm font-medium leading-none">Jackson Lee</p>
-          <p className="text-sm text-muted-foreground">jackson.lee@email.com</p>
-        </div>
-        <div className="ml-auto font-medium">
-          <p className="text-sm font-medium leading-none">Onboarding MyBank</p>
-          <p className="text-sm text-muted-foreground text-end">OMB-199</p>
-        </div>
-      </div>
-      <div className="flex items-center">
-        <Avatar className="h-9 w-9">
-          <AvatarImage src="/avatars/03.png" alt="Avatar" />
-          <AvatarFallback>IN</AvatarFallback>
-        </Avatar>
-        <div className="ml-4 space-y-1">
-          <p className="text-sm font-medium leading-none">Isabella Nguyen</p>
-          <p className="text-sm text-muted-foreground">
-            isabella.nguyen@email.com
-          </p>
-        </div>
-        <div className="ml-auto font-medium">
-          <p className="text-sm font-medium leading-none">Onboarding MyBank</p>
-          <p className="text-sm text-muted-foreground text-end">OMB-198</p>
-        </div>
-      </div>
-      <div className="flex items-center">
-        <Avatar className="h-9 w-9">
-          <AvatarImage src="/avatars/04.png" alt="Avatar" />
-          <AvatarFallback>WK</AvatarFallback>
-        </Avatar>
-        <div className="ml-4 space-y-1">
-          <p className="text-sm font-medium leading-none">William Kim</p>
-          <p className="text-sm text-muted-foreground">will@email.com</p>
-        </div>
-        <div className="ml-auto font-medium">
-          <p className="text-sm font-medium leading-none">Onboarding MyBank</p>
-          <p className="text-sm text-muted-foreground text-end">OMB-197</p>
-        </div>
-      </div>
-      <div className="flex items-center">
-        <Avatar className="h-9 w-9">
-          <AvatarImage src="/avatars/05.png" alt="Avatar" />
-          <AvatarFallback>SD</AvatarFallback>
-        </Avatar>
-        <div className="ml-4 space-y-1">
-          <p className="text-sm font-medium leading-none">Sofia Davis</p>
-          <p className="text-sm text-muted-foreground">sofia.davis@email.com</p>
-        </div>
-        <div className="ml-auto font-medium">
-          <p className="text-sm font-medium leading-none">Onboarding MyBank</p>
-          <p className="text-sm text-muted-foreground text-end">OMB-196</p>
-        </div>
-      </div>
+      {tasks.map((task) => {
+        return (
+          <div className="flex items-center">
+            <Avatar className="h-9 w-9">
+              <AvatarImage src="/avatars/01.png" alt="Avatar" />
+              <AvatarFallback>{task.createdBy.username.slice(0, 2)}</AvatarFallback>
+            </Avatar>
+            <div className="ml-4 space-y-1">
+              <p className="text-sm font-medium leading-none">{task.createdBy.username}</p>
+              <p className="text-sm text-muted-foreground">
+              {task.createdBy.email}
+              </p>
+            </div>
+            <div className="ml-auto font-medium">
+              <p className="text-sm font-medium leading-none">
+                {task.title}
+              </p>
+              <p className="text-sm text-muted-foreground text-end">{task.label}</p>
+            </div>
+          </div>
+        );
+      })}
     </div>
   );
 }
