--- frontend/services/auth/dto/user.ts
@@ -1,6 +1,0 @@
-export interface User {
-  username: string;
-  email: string;
-  id: string;
-  fcmToken: string;
-}
